# stack/apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Administration Stack API",
    "version": "v1",
    "description": "Odoo/staff-app/client-app contract for the Trial Org and Client Org record of truth (docs/adr/0034, docs/adr/0036)."
  },
  "servers": [],
  "components": {
    "securitySchemes": {
      "orgToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque per-org token issued at provision time (docs/adr/0036). Scoped to exactly one org."
      },
      "sigv4": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AWS SigV4 (docs/adr/0036), verified at the front door before the request reaches this service - not a literal API key. Modeled as apiKey here because OpenAPI 3.0 has no native SigV4 scheme."
      }
    },
    "schemas": {
      "OrgId": {
        "type": "string",
        "format": "uuid",
        "description": "The Administration Stack's own identifier for a Trial Org or Client Org."
      },
...
      "OrgRegistration": {
        "type": "object",
        "properties": {
          "orgId": {
            "$ref": "#/components/schemas/OrgId"
          },
          "type": {
            "$ref": "#/components/schemas/OrgType"
          },
          "state": {
            "$ref": "#/components/schemas/OrgState"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "seatsUsed": {
            "type": "integer",
            "minimum": 0
          },
          "seatsTotal": {
            "$ref": "#/components/schemas/SeatsTotal"
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time",
            "description": "Absent for a Client Org, which has no Auto-Destroy expiry."
          }
        },
        "required": [
          "orgId",
          "type",
          "state",
          "name",
          "domain",
          "seatsUsed",
          "seatsTotal"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Stack-specific machine-readable error code, stable across `detail` wording changes."
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
    "/v1/org/{orgId}/registration": {
      "get": {
        "summary": "Read an org's own Org Registration",
        "description": "docs/contexts/hosting/CONTEXT.md: \"the read-only summary of an org's own standing\". The org token must scope to {orgId} itself (docs/adr/0036) - any other {orgId} is 403.",
        "tags": [
          "org"
        ],
        "security": [
          {
            "orgToken": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrgId"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgRegistration"
                }
              }
            }
          },
          "400": {
            "description": "Malformed orgId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid org token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Org token does not scope to this org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orgs/{orgId}": {
      "get": {
        "summary": "Read any Trial Org or Client Org record (staff plane)",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sigv4": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrgId"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgRegistration"
                }
              }
            }
          },
          "400": {
            "description": "Malformed orgId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin principal (SigV4 not verified at the front door)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Change dnsSubdomainLabel - only legal while the org is still `issued`",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sigv4": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrgId"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrgRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "dnsSubdomainLabel already in use, the org has left `issued`, or an Idempotency-Key conflict (reused for a different request, or still processing - see `Retry-After`)",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "description": "Seconds to wait before retrying - only set when the 409 is an Idempotency-Key still-processing conflict.",
                  "example": "3"
                },
                "required": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orgs": {
      "post": {
        "summary": "Create a Trial Org or Client Org, starting in the `issued` state",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sigv4": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "dnsSubdomainLabel already in use, or an Idempotency-Key conflict (reused for a different request, or still processing - see `Retry-After`)",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "description": "Seconds to wait before retrying - only set when the 409 is an Idempotency-Key still-processing conflict.",
                  "example": "3"
                },
                "required": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/orgs/{orgId}/check-status": {
      "post": {
        "summary": "Poll the org's currently-running job to a terminal status, if it has one",
        "description": "Production entry point for `Provisioner.checkStatus` (#298, #281): a safe no-op when the org isn't currently running a job. Reachable per-org by an external scheduler that already knows which orgs have a job in flight.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sigv4": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrgId"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin principal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "An Idempotency-Key conflict (reused for a different request, or still processing - see `Retry-After`)",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "description": "Seconds to wait before retrying - only set when the 409 is an Idempotency-Key still-processing conflict.",
                  "example": "3"
                },
                "required": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  }
}